* scheme-interpreter/table-tests.sml (TableTests.suite): add a suite of tests for the table implementation using two `parallel' lists.
* scheme-interpreter/table.sig (TABLE): introduce a signature for
the table concept and a first implementation of a functor to build
TABLE structures.
(MakeTableWithStringIdentifierAndDoubleListImpl): refined the
implementation of the functor for the table in order to
parameterize the identifier type, instead of setting it to
`string'.

* scheme-interpreter/table.sig: refactored the functor for
parameterizing on the type of `stuff' value too (as done for the
type of identifier).